requirements.txt
keyboard
pystray
Pillow
pyperclip
pyinstaller
